make defaults.device work; normalize tests
